converted rgb to float, to suit PointXYZRGB pcl structure

+++ examples/protonect/include/libfreenect2/registration.h
@@ -50,7 +50,7 @@ public:
   void apply(const Frame* rgb, const Frame* depth, Frame* undistorted, Frame* registered, const bool enable_filter = true, Frame* bigdepth = 0) const;
 
   // compute point XYZ RGB from undistored and registered frames
-  void getPointXYZRGB (const Frame* undistorted, const Frame* registered, int r, int c, float& x, float& y, float& z, unsigned int& rgb) const;
+  void getPointXYZRGB (const Frame* undistorted, const Frame* registered, int r, int c, float& x, float& y, float& z, float& rgb) const;
 
 private:
   void distort(int mx, int my, float& dx, float& dy) const;
